pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-benchmark-suite"
version = "0.1.0"
description = "Production-style benchmarking and evaluation suite for vLLM, TensorRT-LLM, and ONNX Runtime backends."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "OpenAI Codex" }]
dependencies = [
  "click>=8.1.7",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "httpx>=0.27.0",
  "rich>=13.7.1",
  "jinja2>=3.1.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
]
semantic = [
  "sentence-transformers>=3.0.1",
]

[project.scripts]
benchmark-suite = "llm_benchmark_suite.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
